Code Actions Support (Ask Cody to Fix) (#1921)
Add the ability for the Agent to provide CodeActions. In practice this
means "Ask Cody to Fix" is now supported as a QuickFix on errors shown
in JetBrains.

https://github.com/user-attachments/assets/f3bad3fd-cd67-427c-baa6-449a4c0dfd48

## FAQ

**Q: What errors are supported?**
A: Any errors shown in the JetBrains UI. Cody doesn't do any code
analysis, we simply pass the issues that JetBrains is showing along.
These Issues (or Inspections as they're known) can come both from
JetBrain's language support, external Plugins as well as custom
Inspections.

**Q: Can Cody explain the error like in VSCode?**
A: Not yet, but the foundational support is there now so it will be easy
enough to add soon.

Fixes CODY-2788, CODY-2830

## Test plan
- Verified that existing unit & integration tests work with Protocol
migrated protocol classes
- Manually verified that "Ask Cody to Fix" command is correctly shown
and invoked. Most of the logic remains on the Agent.
- Isolated changes on FixupSession to FixupSessionV2 to not introduce
bugs in FixupSessionV1 dependent codepaths which currently have low test
coverage.

+++ src/main/kotlin/com/sourcegraph/cody/autocomplete/CodyAutocompleteManager.kt
@@ -34,11 +34,12 @@ import com.sourcegraph.cody.agent.protocol.AutocompleteTriggerKind
 import com.sourcegraph.cody.agent.protocol.CompletionItemParams
 import com.sourcegraph.cody.agent.protocol.ErrorCode
 import com.sourcegraph.cody.agent.protocol.ErrorCodeUtils.toErrorCode
-import com.sourcegraph.cody.agent.protocol.Position
 import com.sourcegraph.cody.agent.protocol.ProtocolTextDocument.Companion.uriFor
-import com.sourcegraph.cody.agent.protocol.Range
 import com.sourcegraph.cody.agent.protocol.RateLimitError.Companion.toRateLimitError
 import com.sourcegraph.cody.agent.protocol.SelectedCompletionInfo
+import com.sourcegraph.cody.agent.protocol_extensions.Position
+import com.sourcegraph.cody.agent.protocol_generated.Position
+import com.sourcegraph.cody.agent.protocol_generated.Range
 import com.sourcegraph.cody.autocomplete.render.AutocompleteRendererType
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteBlockElementRenderer
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteElementRenderer
@@ -238,7 +239,7 @@ class CodyAutocompleteManager {
                 SelectedCompletionInfo(
                     lookupString,
                     if (startPosition < 0) Range(position, position)
-                    else Range(Position(lineNumber.toLong(), startPosition.toLong()), position)))
+                    else Range(Position(lineNumber, startPosition), position)))
     notifyApplication(project, CodyStatus.AutocompleteInProgress)
 
     val resultOuter = CompletableFuture<Void?>()
